Split each mode into its own src/test files
