+++ lib/webofthings/td-json-schema-validation.json
