+++ lib/webofthings/wotutils.js
